requirements.txt
# Twitch Chat Collector - 依存パッケージ

# Twitch API クライアント
requests>=2.31.0

# WebSocket通信
websockets>=12.0

# データベース
SQLAlchemy>=2.0.0

# 環境変数管理
python-dotenv>=1.0.0

# 設定ファイル読み込み
PyYAML>=6.0
